+++ R/encoding.R
@@ -77,7 +77,7 @@ as.utf8 <- function(x, from){
 #' @rdname encodings
 as.nativeEnc <- function(x, from){
   Encoding(x) <- from
-  y <- enc2native(x)
+  y <- enc2native(x) # has been benchmarked, is faster than iconv and stringi::stri_encode
   Encoding(y) <- "unknown"
   y
 }